Refactor Rabbit and Redis broker implementations to include logger support and streamline connection management
